rename `client/file_storage/classes/main.py` to `client/file_storage/classes/interface.py` and move `LocalStorageClient` into it's own file
