+++ core/clients/file_storage/classes/local_storage.py
@@ -1,29 +1,7 @@
-from abc import ABC, abstractmethod
 import pathlib
 import os
 
-class FileStorageClient(ABC):
-    PREFIX: str
-
-    @abstractmethod
-    def __init__(self, path: str):
-        raise NotImplementedError
-    
-    @abstractmethod
-    def file_path(self, file_name: str) -> str:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_file(self, file_name: str) -> bytes:
-        raise NotImplementedError
-
-    @abstractmethod
-    def put_file(self, file_name: str, contents: bytes) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
-    def delete_file(self, file_name: str) -> None:
-        raise NotImplementedError
+from dev.yeti.core.clients.file_storage.classes.interface import FileStorageClient
 
 class LocalStorageClient(FileStorageClient):
     PREFIX = ""
